Extract _get_file_path to s3._write
